tools/scraper/requirements.txt
duckduckgo-search>=6.0
python-dotenv>=1.0
